4.22

+++ controllers/blogs.js
@@ -1,6 +1,8 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+
 
 
 blogsRouter.get('/', async (request, response) => {
@@ -11,11 +13,21 @@ blogsRouter.get('/', async (request, response) => {
   
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
-  const user = await User.findById(body.user)
+  if (!request.token){
+    return response.status(401).json({ error: 'token missing' })
+  }
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
 
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token  invalid' })
+  }
+
+  const user = await User.findById(decodedToken.id)
+  console.log(decodedToken)
   if (body.title === undefined || body.url === undefined) {
     return response.status(400).json({ error: 'title or url missing' })
   }
+
   const blog = new Blog({
     title: body.title,
     author: body.author,
@@ -23,14 +35,34 @@ blogsRouter.post('/', async (request, response) => {
     likes: body.likes,
     user: user.id
   })
+
   const savedBlog = await blog.save()
+
   user.blogs = user.blogs.concat(savedBlog._id)
   await user.save()
+  response.json(savedBlog.toJSON())
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  if (!request.token){
+    return response.status(401).json({ error: 'token missing' })
+  }
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+  console.log(blog)
+  if ( blog.user.toString() === user.id.toString() ){
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  }
+
+
 })
 
 blogsRouter.put('/:id', (request, response, next) => {
